Added new test function -> modified main.py and slice_segthor.py
Added a way to test our newly trained models and compare to the benchmark. Had to change slice_segthor to slice the dataset properly. Kept patients 7, 17, 18, 34 and 39 as the testset which training never sees. Original code should work without issues as new flags added to use this functionality. Check arguments of both main.py and slice_segthor.py to see